+++ src/main/java/com/github/qeaml/mmic/mixin/OptionsScreenMixin.java
@@ -5,16 +5,16 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.github.qeaml.mmic.PlaytimeScreen;
+import com.github.qeaml.mmic.gui.MenuScreen;
 
 import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.TitleScreen;
+import net.minecraft.client.gui.screen.option.OptionsScreen;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.text.Text;
 
-@Mixin(TitleScreen.class)
-public abstract class TitleScreenMixin extends Screen {
-  protected TitleScreenMixin(Text title) {
+@Mixin(OptionsScreen.class)
+public abstract class OptionsScreenMixin extends Screen {
+  protected OptionsScreenMixin(Text title) {
     super(title);
   }
 
@@ -23,8 +23,12 @@ public abstract class TitleScreenMixin extends Screen {
     at = @At("HEAD")
   )
   private void init(CallbackInfo ci) {
-    addDrawableChild(new ButtonWidget(20, height - 40, 98, 20, Text.translatable("gui.mmic.sessions"), button -> {
-      client.setScreen(new PlaytimeScreen(this));
+    addDrawableChild(new ButtonWidget(
+      10, height - 30,
+      60, 20,
+      Text.translatable("gui.mmic.menu"),
+    (button) -> {
+      client.setScreen(new MenuScreen(this));
     }));
   }
 }